baremetal: integrate provisioning logic
The bare-metal platform lacked any understanding of whether an instance
is actually running the configuration that lokoctl put into matchbox,
because, when the configuration was updated, there was no notification
to the user that PXE booting has to be done again for this instance.
Also, when to boot from PXE was also clear to the user because it must
happen after lokoctl populated matchbox with the (new) configuration
but before any other steps are timing out.
A null resource and is used with the possibility to run user-defined
commands that can automate PXE provisioning at the right time, i.e.,
initially at the first lokoctl run and after each Ignition
configuration change. However, PXE booting is a long and maybe even
manual process, or even impossible at production due to the lack of
an appropriate DHCP server. We can rely on Ignition to simulate
reprovisioning by creating the first_boot flag file via SSH and
issuing a reboot, which makes Ignition fetch the configuration from
matchbox, and if we make sure to clean the root filesystem the result
is the same as if reprovisioned with a PXE boot.